sse assgn/assgn5 fortify phase/requirements.txt
pyelftools
